Migrate all entity IDs from Guid to long (bigint)
BaseEntity.Id: Guid -> long (auto-increment)
All related entities, DTOs, controllers, services, and migrations updated.
Old migrations deleted, replaced with single InitialCreate migration.

This was already applied to the running DB but never committed to git,
causing CI to fail with 'Guid == long' type mismatch errors.

+++ src/VypusknykPlus.Application/DTOs/Cart/AddCartItemRequest.cs
@@ -4,7 +4,7 @@ namespace VypusknykPlus.Application.DTOs.Cart;
 
 public class AddCartItemRequest
 {
-    public int? ProductId { get; set; }
+    public long? ProductId { get; set; }
     public string? Name { get; set; }
     public decimal? Price { get; set; }
     public int Qty { get; set; }
